Bug 1210109 - Make home panel buttons the right size             - Make the menu button the right colour and positioned correctly             - separate the menu from the specific Browser implementation of it             - Make new custom UIBarButtonItems tappable
