+++ Client/Frontend/Menu/AppMenuConfiguration.swift
@@ -4,37 +4,6 @@
 
 import Foundation
 
-protocol MenuConfiguration {
-
-    func menuForState(appState: AppState) -> MenuConfiguration
-
-    var menuItems: [MenuItem] { get }
-    var menuToolbarItems: [MenuToolbarItem]? { get }
-    var numberOfItemsInRow: Int { get }
-
-    init(appState: AppState)
-    func toolbarColor() -> UIColor
-    func toolbarTintColor() -> UIColor
-    func menuBackgroundColor() -> UIColor
-    func menuTintColor() -> UIColor
-    func menuFont() -> UIFont
-    func menuIcon() -> UIImage?
-    func shadowColor() -> UIColor
-    func selectedItemTintColor() -> UIColor
-}
-
-protocol MenuActionDelegate: class {
-    func performMenuAction(action: MenuAction, withAppState appState: AppState)
-}
-
-struct MenuAction {
-    let action: String!
-
-    init(action: String) {
-        self.action = action
-    }
-}
-
 enum AppMenuAction: String {
     case OpenNewNormalTab = "OpenNewNormalTab"
     case OpenNewPrivateTab = "OpenNewPrivateTab"
