+++ Client/Frontend/Widgets/Menu/MenuDelegate.swift
